commenting out a small section of the `createNonPinnedSameMajorMinorPatchVersion` method.
For the moment, I don't remember why it was there.  It does not seem to serve an purpose.  The only thing that I can think of is that it somehow would prime the working lists in some manner that I don't recognize yet.  If that is the case, then commenting this out should produce odd behavior and I can recognize what this was all about again...... and then I will comment what that is!!!!  ;-)